+++ js/about.js
@@ -1,0 +1,12 @@
+import { burgerCon } from './modules/burger-menu.js';
+// import { contactForm } from './modules/contact-form.js';
+import { gsapGlobal } from './modules/gsap-global.js';
+import { gsapContactScroll } from './modules/gsap-scroll-contact.js';
+import { gsapAbout} from './modules/gsap-about.js';
+
+
+burgerCon();
+// contactForm();
+gsapGlobal();
+gsapContactScroll();
+gsapAbout();
