+++ js/case-study.js
@@ -1,0 +1,14 @@
+import { burgerCon } from './modules/burger-menu.js';
+import { slider } from './modules/slider-animation.js';
+import { videoPlayer } from './modules/video-player.js';
+import { gsapGlobal } from './modules/gsap-global.js';
+import { gsapCards } from './modules/gsap-cards.js';
+import { gsapCaseStudy } from './modules/gsap-cs.js';
+
+
+burgerCon();
+slider();
+videoPlayer();
+gsapGlobal();
+gsapCards();
+gsapCaseStudy();
