+++ js/home.js
@@ -1,0 +1,16 @@
+import { burgerCon } from './modules/burger-menu.js';
+import { contactForm } from './modules/contact-form.js';
+import { videoPlayer } from './modules/video-player.js';
+import { gsapGlobal } from './modules/gsap-global.js';
+import { gsapHome } from './modules/gsap-home.js';
+import { gsapContactScroll } from './modules/gsap-scroll-contact.js';
+import { gsapCards } from './modules/gsap-cards.js';
+
+
+burgerCon();
+contactForm();
+videoPlayer();
+gsapGlobal();
+gsapHome();
+gsapCards();
+gsapContactScroll();
