--- js/main.js
@@ -1,12 +1,0 @@
-import { initializeVideoPlayer } from './video-player.js';
-import { initializeHamburgerMenu } from './burger-menu.js';
-import { initializeGSAPAnimations } from './gsap-animation.js';
-import { initializeSlider } from './slider-animation.js';
-import { contactForm } from './contact-form.js';
-
-// Call your functions
-initializeVideoPlayer();
-initializeHamburgerMenu();
-initializeGSAPAnimations();
-initializeSlider();
-contactForm();
