+++ js/projects.js
@@ -1,0 +1,11 @@
+import { burgerCon } from './modules/burger-menu.js';
+import { videoPlayer } from './modules/video-player.js';
+import { gsapGlobal } from './modules/gsap-global.js';
+import { gsapProjects } from './modules/gsap-project.js';
+
+
+
+burgerCon();
+videoPlayer();
+gsapGlobal();
+gsapProjects();
