+++ sass/abstracts/_variables.scss
@@ -3,3 +3,6 @@ $secondary-color: #6933B3;
 $background-color: #09060C;
 $font-family: 'Inter', sans-serif;
 $heading-font-weight: bold;
+
+
+//#8957E5 - new purple, maybe
